+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/object/SMGRegion.java
@@ -17,11 +17,11 @@ import org.sosy_lab.cpachecker.cpa.smg.graphs.value.SMGValue;
 @Immutable
 public final class SMGRegion extends SMGObject implements SMGObjectTemplate {
 
-  public SMGRegion(int pSize, String pLabel) {
+  public SMGRegion(long pSize, String pLabel) {
     super(pSize, pLabel, SMGObjectKind.REG);
   }
 
-  public SMGRegion(int pSize) {
+  public SMGRegion(long pSize) {
     super(pSize, "ID" + SMGCPA.getNewValue(), SMGObjectKind.REG);
   }
 
@@ -29,13 +29,13 @@ public final class SMGRegion extends SMGObject implements SMGObjectTemplate {
     super(pOther);
   }
 
-  public SMGRegion(int pSize, String pLabel, int pLevel) {
+  public SMGRegion(long pSize, String pLabel, int pLevel) {
     super(pSize, pLabel, pLevel, SMGObjectKind.REG);
   }
 
   @Override
   public String toString() {
-    return "REGION("+ getLabel() + ", " + getSize() + "b, " + "level=" + getLevel() + ")";
+    return "REGION(" + getLabel() + ", " + getSize() + "b, " + "level=" + getLevel() + ")";
   }
 
   @Override
